requirements.txt
streamlit==1.31.0
youtube-transcript-api==0.6.2
google-generative-ai==0.4.1
